+++ sample-app/app/mailboxes/application_mailbox.rb
@@ -1,3 +1,3 @@
 class ApplicationMailbox < ActionMailbox::Base
-  # routing /something/i => :somewhere
+  routing /.*/ => :test
 end